requirements.txt.txt
streamlit
pandas
google-generativeai
fpdf
Pillow
openpyxl